Move SkeletonUtils functinos to Skeleton ;
