+++ Plugins/Animation/src/Drawing/SkeletonBoneDrawable.cpp
@@ -6,8 +6,6 @@
 #include <Engine/Renderer/RenderObject/RenderObjectManager.hpp>
 #include <Engine/Renderer/RenderTechnique/ShaderConfigFactory.hpp>
 
-#include <Core/Animation/Handle/SkeletonUtils.hpp>
-
 namespace AnimationPlugin {
 
 SkeletonBoneRenderObject::SkeletonBoneRenderObject( const std::string& name,
@@ -55,7 +53,7 @@ void SkeletonBoneRenderObject::update() {
 void SkeletonBoneRenderObject::updateLocalTransform() {
     Ra::Core::Vector3 start;
     Ra::Core::Vector3 end;
-    Ra::Core::Animation::SkeletonUtils::getBonePoints( m_skel, m_id, start, end );
+    m_skel.getBonePoints( m_id, start, end );
 
     Ra::Core::Transform scale = Ra::Core::Transform::Identity();
     scale.scale( ( end - start ).norm() );
