+++ src/Core/Animation/Handle/Skeleton.hpp
@@ -30,20 +30,19 @@ class RA_CORE_API Skeleton : public PointCloud {
 
     /// SIZE
     inline uint size() const override { return m_graph.size(); }
-    virtual void clear() override; // Empty the data of the skeleton
+    void clear() override; // Empty the data of the skeleton
 
     /// SPACE INTERFACE
-    virtual const Pose&
-    getPose( const SpaceType MODE ) const override; // Return the pose in MODE space
-    virtual void setPose( const Pose& pose,
-                          const SpaceType MODE ) override; // Set the MODE space pose
-    virtual const Transform& getTransform( const uint i, const SpaceType MODE )
-        const override; // Return the i-th transform matrix of the pose in MODE space
-    virtual void
-    setTransform( const uint i, const Transform& T,
-                  const SpaceType MODE ) override; // Set the i-th transform of the MODE space pose
+    const Pose& getPose( const SpaceType MODE ) const override; // Return the pose in MODE space
+    void setPose( const Pose& pose, const SpaceType MODE ) override; // Set the MODE space pose
+    const Transform& getTransform( const uint i, const SpaceType MODE ) const override; // Return the i-th transform matrix of the pose in MODE space
+    void setTransform( const uint i, const Transform& T, const SpaceType MODE ) override; // Set the i-th transform of the MODE space pose
 
-    void getBonePoints( uint i, Vector3& startOut, Vector3& endOut ) const;
+    /// UTILS
+    void getBonePoints( uint i, Vector3& startOut, Vector3& endOut ) const; // Return the end points of the i-th bone
+    Ra::Core::Vector3 projectOnBone( int boneIdx, const Ra::Core::Vector3& pos ) const; // Returns the closest point of the i-th bone
+
+    friend std::ostream& operator<<( std::ostream& os, const Skeleton& skeleton ); // prints the skeleton hierarchy to stream os
 
     /// VARIABLE
     Ra::Core::Graph::AdjacencyList m_graph; // The adjacency list.
